Updated API verion back to 1 and used user_v1
